# Conflicts:
#	plugins/fabrik_element/databasejoin/databasejoin.php

+++ plugins/fabrik_element/databasejoin/databasejoin.php
@@ -111,6 +111,11 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	protected $orderBy = '';
 
 	/**
+	 * @var bool
+	 */
+	public $_rawFilter = false;
+
+	/**
 	 * Create the SQL select 'name AS alias' segment for list/form queries
 	 *
 	 * @param   array &$aFields   array of element names
@@ -1057,8 +1062,16 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	{
 		$params         = $this->getParams();
 		$element        = $this->getElement();
+		$filterWhere    = trim($params->get('database_join_filter_where_sql', ''));
 		$whereAccess    = $params->get('database_join_where_access', 26);
-		$where          = ($this->mustApplyWhere($whereAccess, $element->id) && $incWhere) ? $params->get('database_join_where_sql') : '';
+		$where          = '';
+
+		// $$$ hugh - 5/13/2019 - filter where now overrides form where, if present
+		if (FArrayHelper::getValue($opts, 'mode', '') !== 'filter' || empty($filterWhere))
+		{
+			$where = ($this->mustApplyWhere($whereAccess, $element->id) && $incWhere) ? $params->get('database_join_where_sql') : '';
+		}
+
 		$join           = $this->getJoin();
 		$thisTableAlias = is_null($thisTableAlias) ? $join->table_join_alias : $thisTableAlias;
 		$repeatCounter  = FArrayHelper::getValue($opts, 'repeatCounter', 0);
@@ -1091,7 +1104,6 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		 * but first cut it's safer to put it here (don't ask).
 		 */
 
-		$filterWhere = trim($params->get('database_join_filter_where_sql', ''));
 		if (FArrayHelper::getValue($opts, 'mode', '') === 'filter' && !empty($filterWhere))
 		{
 			if (preg_match('/(ORDER\s+BY)(.*)/is', $filterWhere, $matches))
@@ -1119,7 +1131,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		}
 
 		$where = $w->parseMessageForRepeats($where, $data, $this, $repeatCounter);
-		$where = $w->parseMessageForPlaceHolder($where, $data, false);
+		$where = $w->parseMessageForPlaceHolder($where, $data, false, false, null, false);
 
 		if (!$query)
 		{
@@ -1567,7 +1579,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			return false;
 		}
 
-		$db    = $this->getDb();
+		$db = FabrikWorker::getDbo(true);
 		$query = $db->getQuery(true);
 		$query->select('id')->from('#__{package}_lists')->where('form_id =' . $popupFormId);
 		$db->setQuery($query);
@@ -2448,7 +2460,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		}
 		else
 		{
-			if (!preg_match('/^WHERE\s+/i', $where))
+			if (!preg_match('/^\s*WHERE\s+/i', $where))
 			{
 				$where = 'WHERE ' . $where;
 			}
@@ -2725,11 +2737,12 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		else
 		{
 			$group = $this->getGroup();
+			$exactMatch = $this->getElement()->get('filter_exact_match');
 
 			if (!$group->isJoin() && $group->canRepeat())
 			{
 				// Deprecated I think - repeat groups are always joins.
-				$fval = $this->getElement()->get('filter_exact_match') ? $originalValue : $value;
+				$fval = $exactMatch ? $originalValue : $value;
 				$str  = " ($key = $fval OR $key LIKE \"$originalValue',%\"" . " OR $key LIKE \"%:'$originalValue',%\""
 					. " OR $key LIKE \"%:'$originalValue'\"" . " )";
 			}
@@ -2750,11 +2763,40 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 					}
 					$parentID	= $this->parentID();
 
-					$groupBy = $this->_db->qn($dbName . '.' . $parentID);
-					$rows    = $this->checkboxRows($groupBy, $condition, $value, $where);
-					$joinIds = array_keys($rows);
+					if (($fType === 'multiselect' || $fType === 'checkbox') && $exactMatch)
+					{
+						$rows    = $this->checkboxRows('id', $condition, $value, $where);
+						$joinIds = array();
+						$matches = array();
 
-					if (!empty($rows))
+						foreach ($rows as $row)
+						{
+							if (!array_key_exists($row->parent_id, $matches))
+							{
+								$matches[$row->parent_id] = array();
+							}
+
+							$matches[$row->parent_id][] = $row->value;
+						}
+
+						foreach ($matches as $joinId => $match)
+						{
+							$diff = array_diff($originalValue, $match);
+
+							if (empty($diff))
+							{
+								$joinIds[] = $joinId;
+							}
+						}
+					}
+					else
+					{
+						$groupBy = $this->_db->qn($dbName . '.' . $parentID);
+						$rows    = $this->checkboxRows($groupBy, $condition, $value, $where);
+						$joinIds = array_keys($rows);
+					}
+
+					if (!empty($joinIds))
 					{
 						// Either look for the ids in the main fabrik list or the group's list.
 						$groupJoinModel = $group->getJoinModel();
@@ -3075,7 +3117,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		$modalId                  = 'dbjoin_popupform';
 		$opts                     = $this->_getOptionVals();
 		$data                     = $this->getFormModel()->data;
-		$arSelected               = $this->getValue($data, $repeatCounter);
+		$arSelected               = $this->getValue($data, $repeatCounter, array('raw' => true));
 		$table                    = $params->get('join_db_name');
 		$opts                     = $this->getElementJSOptions($repeatCounter);
 		$forms                    = $this->getLinkedForms();
@@ -3655,7 +3697,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	private function _autocompleteWhere($how, $field, $search)
 	{
 		$db     = FabrikWorker::getDbo();
-		$search = strtolower($search);
+		$search = mb_strtolower($search);
 		$field  = 'LOWER(' . $field . ')';
 
 		switch ($how)
@@ -4110,9 +4152,21 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			$key = $db->q($key);
 		});
 
-		$query = $db->getQuery(true);
-		$query->delete($db->qn($join->table_join))->where('id IN (' . implode(',', $keys) .')');
+		if (!empty($keys))
+		{
+			$query = $db->getQuery(true);
+			$query->delete($db->qn($join->table_join))->where('id IN (' . implode(',', $keys) . ')');
 
-		return $db->setQuery($query)->execute();
+			try
+			{
+				$db->setQuery($query)->execute();
+			}
+			catch (Exception $e)
+			{
+				// meh
+			}
+		}
+
+		return true;
 	}
 }
